+++ tests/test_formats_id3v2.py
@@ -27,28 +27,28 @@ def _():
 		ID3v2Frames(id3_version=None)
 
 	with raises(ValueError):
-		ID3v2Frames.load(b'data', (1, 0))
+		ID3v2Frames.parse(b'data', (1, 0))
 
 	with raises(ValueError):
-		ID3v2Frames.load(b'data', None)
+		ID3v2Frames.parse(b'data', None)
 
 	v22_frames_init = ID3v2Frames(id3_version=ID3Version.v22)
 	v22_frames_init_tuple = ID3v2Frames(id3_version=(2, 2))
-	v22_frames_load = ID3v2Frames.load(
+	v22_frames_load = ID3v2Frames.parse(
 		(Path(__file__).parent / 'files' / 'audio' / 'test-mp3-id3v22.mp3').read_bytes()[10:],
 		ID3Version.v22,
 	)
 
 	v23_frames_init = ID3v2Frames(id3_version=ID3Version.v23)
 	v23_frames_init_tuple = ID3v2Frames(id3_version=(2, 3))
-	v23_frames_load = ID3v2Frames.load(
+	v23_frames_load = ID3v2Frames.parse(
 		(Path(__file__).parent / 'files' / 'audio' / 'test-mp3-id3v23.mp3').read_bytes()[10:],
 		ID3Version.v23,
 	)
 
 	v24_frames_init = ID3v2Frames(id3_version=ID3Version.v24)
 	v24_frames_init_tuple = ID3v2Frames(id3_version=(2, 4))
-	v24_frames_load = ID3v2Frames.load(
+	v24_frames_load = ID3v2Frames.parse(
 		(Path(__file__).parent / 'files' / 'audio' / 'test-mp3-id3v24.mp3').read_bytes()[10:],
 		ID3Version.v24,
 	)
@@ -115,7 +115,7 @@ def _(
 	tags=['unit', 'id3', 'id3v2', 'ID3v2Header'],
 )
 def test_ID3v2Header():
-	v24_header_load = ID3v2Header.load(Path(__file__).parent / 'files' / 'audio' / 'test-mp3-id3v24.mp3')
+	v24_header_load = ID3v2Header.parse(Path(__file__).parent / 'files' / 'audio' / 'test-mp3-id3v24.mp3')
 	v24_header_init = ID3v2Header(
 		size=2254,
 		flags=ID3v2Flags(
@@ -129,7 +129,7 @@ def test_ID3v2Header():
 
 	assert v24_header_load == v24_header_init
 
-	v24_header_load = ID3v2Header.load(Path(__file__).parent / 'files' / 'audio' / 'test-mp3-id3v24-unsync.mp3')
+	v24_header_load = ID3v2Header.parse(Path(__file__).parent / 'files' / 'audio' / 'test-mp3-id3v24-unsync.mp3')
 	v24_header_init = ID3v2Header(
 		size=2254,
 		flags=ID3v2Flags(
@@ -144,6 +144,6 @@ def test_ID3v2Header():
 	assert v24_header_load == v24_header_init
 
 	with raises(InvalidHeader):
-		ID3v2Header.load(
+		ID3v2Header.parse(
 			(Path(__file__).parent / 'files' / 'audio' / 'test-mp3-id3v24.mp3').read_bytes()[3:]
 		)